Compiler Error Fixes
Fixed A LOT of compiler errors. I recommend Tim going over the changes
to make sure they made sense and followed the original...

Currently, makefile doesn't work. Doesn't compile everything together,
but instead tries to do them separately, resulting in an issue with
"Undefined symbols for architecture x86_64: _main". (It's not finding
the main function in main.cpp because it's in a different directory.)