Game Loop
Added game loop and registered the input controller callbacks with
GLUT. Except now, they’re not being processed. This may be because of a
GLUT mainloop() issue, as it’s never called. More work will need to be
done to resolve this. Possible solution would be to use GLFW instead.